+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Fixed some cases of iterators closing in `Iterator.{ zip, zipKeyed }` polyfills
 - Fixed validation of iterators `.next()` results an objects in `Iterator.{ zip, zipKeyed }` polyfills
 - Fixed a lack of early error in `Iterator.concat` polyfill on primitive as an iterator
+- Fixed buffer mutation exposure in `Iterator.prototype.windows` polyfill
 - Fixed iterator closing in `Set.prototype.{ isDisjointFrom, isSupersetOf }` polyfill
 - Fixed (updated following the final spec) one more case `Set.prototype.difference` polyfill with updating `this`
 - Fixed `DataView.prototype.setFloat16` polyfill in (0, 1) range